# api/openapi.yaml
openapi: 3.1.0
info:
  title: leani native API
  version: 1.0.0
  description: |
    Coverage-aware queries and durable processor changes. All cursors are
    opaque. Ethereum quantities that can exceed JavaScript's safe integer
    range are encoded as unsigned decimal strings.

    Every route answers a `Host` name other than `localhost` and those in
    `api.allowed_hosts` with 421 `host_not_allowed` (IP addresses pass), and
    a browser `Origin` other than a loopback one or one in
    `api.allowed_origins` with 403 `origin_not_allowed`. A GET or HEAD
    without an `Origin` whose `Sec-Fetch-Site` is `cross-site` or
    `same-site` and whose `Sec-Fetch-Mode` is not `navigate` gets 403
    `cross_site_request`; on the two consumer stream routes, so does one in
    any mode, navigations included. A POST or DELETE needs
    `content-type: application/json` or `x-leani-request: 1`, and gets 415
    `unsupported_media_type` otherwise.

    A consumer created with a credential needs it in
    `x-leani-consumer-credential` on every consumer-scoped delivery route,
    in addition to any bearer token, and gets 403 `forbidden` without it.
    Listing, reading, and revoking consumers are administrative: only the
    bearer token protects them, whether or not a consumer has a credential.
servers:
  - url: http://127.0.0.1:8080
security:
  - {}
  - bearerAuth: []
paths:
...
  /v1/q/blobs/blocks:
    get:
      summary: List indexed blob blocks
      operationId: listBlobBlocks
      parameters:
        - name: fromBlock
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/SafeInteger"
        - name: toBlock
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/SafeInteger"
        - $ref: "#/components/parameters/Limit"
        - name: cursor
          in: query
          schema:
            $ref: "#/components/schemas/Cursor"
        - name: allowPartial
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Ordered blocks and explicit requested-range coverage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobBlockPage"
        "409":
          description: The range contains unindexed gaps.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        default:
          $ref: "#/components/responses/Error"
  /v1/q/blobs/blocks/{number}:
    parameters:
      - name: number
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/SafeInteger"
    get:
      summary: Read an indexed blob block
      operationId: getBlobBlock
      responses:
        "200":
          description: One indexed blob block.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobBlock"
        default:
          $ref: "#/components/responses/Error"
  /v1/q/blobs/snapshot:
    get:
      summary: Read an atomic blob block snapshot
      operationId: listBlobSnapshot
      parameters:
        - name: fromBlock
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/SafeInteger"
        - name: toBlock
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/SafeInteger"
        - $ref: "#/components/parameters/Limit"
        - name: cursor
          in: query
          schema:
            $ref: "#/components/schemas/Cursor"
        - name: allowPartial
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: |
            Ordered blob blocks with their complete blob-transaction children,
            suitable for bulk materialized-view bootstrapping.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobSnapshotPage"
        "409":
          description: The range contains unindexed gaps.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        default:
          $ref: "#/components/responses/Error"
  /v1/q/blobs/transactions:
    get:
      summary: List indexed blob transactions
      operationId: listBlobTransactions
      parameters:
        - name: blockNumber
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/SafeInteger"
        - $ref: "#/components/parameters/Limit"
        - name: cursor
          in: query
          schema:
            $ref: "#/components/schemas/Cursor"
      responses:
        "200":
          description: Blob transactions in canonical transaction order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobTransactionPage"
        default:
          $ref: "#/components/responses/Error"
  /v1/q/blobs/transactions/{hash}:
    parameters:
      - name: hash
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Hash32"
    get:
      summary: Read an indexed blob transaction
      operationId: getBlobTransaction
      responses:
        "200":
          description: One indexed blob transaction.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobTransaction"
        default:
          $ref: "#/components/responses/Error"
  /v1/q/blobs/schedule:
    get:
      summary: Read the active blob schedule
      operationId: getBlobSchedule
      parameters:
        - name: atBlock
          in: query
          schema:
            $ref: "#/components/schemas/SafeInteger"
      responses:
        "200":
          description: Active known blob schedule entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobSchedule"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
...
  responses:
    Error:
      description: Stable machine-readable error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    SafeInteger:
      type: integer
      minimum: 0
      maximum: 9007199254740991
    DecimalQuantity:
      type: string
      pattern: "^(0|[1-9][0-9]*)$"
      examples: ["1000000000000000000"]
...
    Hash32:
      type: string
      pattern: "^0x[0-9a-f]{64}$"
    Address:
      type: string
      pattern: "^0x[0-9a-f]{40}$"
    Cursor:
      type: string
      minLength: 1
      description: Store-, chain-, and processor-bound opaque value.
    Finality:
      type: string
      enum: [preview, included, finalized]
      description: |
        Chain-confidence status of the block a record belongs to.
        `included`: the block is on the currently followed chain; a reorg can
        still remove it. `finalized`: Ethereum consensus has finalized the
        block, as verified by this node's configured finality source. When a
        source serves finalized material under dataset trust (Xatu, EraE), the
        label reflects that source's trust model, not an independent proof.
        `preview`: a peer supplied the block and its connection to the chain is
        still being checked; only the embedded CLI emits `preview`, the HTTP
        API never does. A record's `data.finality`, when present, always equals
        the envelope `finality`.
...
    CoverageInterval:
      type: object
      description: >-
        Inclusive covered range whose finality is uniform for every block in
        the interval. Ranges are split at coverage gaps and exact finality
        boundaries.
      additionalProperties: false
      required: [fromBlock, toBlock, finality]
      properties:
        fromBlock:
          $ref: "#/components/schemas/SafeInteger"
        toBlock:
          $ref: "#/components/schemas/SafeInteger"
        finality:
          $ref: "#/components/schemas/Finality"
    RequestedRange:
      type: object
      additionalProperties: false
      required: [fromBlock, toBlock]
      properties:
        fromBlock:
          $ref: "#/components/schemas/SafeInteger"
        toBlock:
          $ref: "#/components/schemas/SafeInteger"
    ProcessorCoverage:
      type: object
      additionalProperties: false
      required:
        [chainId, chainFinalizedHead, available, configuredStartBlock, processedThrough, finalizedThrough, complete, state]
      properties:
        chainId:
          $ref: "#/components/schemas/SafeInteger"
        chainFinalizedHead:
          oneOf:
            - $ref: "#/components/schemas/ChainFinalizedHead"
            - type: "null"
        requested:
          oneOf:
            - $ref: "#/components/schemas/RequestedRange"
            - type: "null"
        available:
          type: array
          description: >-
            Exact processor coverage split at gaps and finality boundaries.
            Use these intervals when selecting finalized coverage.
          items:
            $ref: "#/components/schemas/CoverageInterval"
        configuredStartBlock:
          $ref: "#/components/schemas/SafeInteger"
        processedThrough:
          oneOf:
            - $ref: "#/components/schemas/SafeInteger"
            - type: "null"
        finalizedThrough:
          description: >-
            Highest covered block recorded as finalized. This is not a
            contiguous-coverage guarantee; lower gaps or non-finalized blocks
            can exist. Inspect available intervals for exact finalized
            coverage.
          oneOf:
            - $ref: "#/components/schemas/SafeInteger"
            - type: "null"
        complete:
          type: boolean
        state:
          description: |
            `paused` and `failed` report the processor's live lane;
            `/v1/network/status` gives its `pauseReason`.
          type: string
          enum: [starting, backfilling, catching_up, live, degraded, paused, failed]
    ChainFinalizedHead:
      type: object
      additionalProperties: false
      required: [number, hash]
      properties:
        number:
          $ref: "#/components/schemas/SafeInteger"
        hash:
          $ref: "#/components/schemas/Hash32"
...
    BlobBlock:
      type: object
      additionalProperties: false
      required:
        - network
        - blockNumber
        - blockHash
        - parentHash
        - timestamp
        - size
        - blobCount
        - blobGasUsed
        - excessBlobGas
        - blobBaseFee
        - executionBaseFee
        - gasUsed
        - gasLimit
        - executionEthBurnedWei
        - blobEthBurnedWei
        - reserveFeeWei
        - transactionCount
        - targetBlobsPerBlock
        - maxBlobsPerBlock
        - transformVersion
        - finality
      properties:
        network:
          type: string
        blockNumber:
          $ref: "#/components/schemas/SafeInteger"
        blockHash:
          $ref: "#/components/schemas/Hash32"
        parentHash:
          $ref: "#/components/schemas/Hash32"
        timestamp:
          $ref: "#/components/schemas/SafeInteger"
        size:
          $ref: "#/components/schemas/DecimalQuantity"
        blobCount:
          $ref: "#/components/schemas/SafeInteger"
        blobGasUsed:
          $ref: "#/components/schemas/DecimalQuantity"
        excessBlobGas:
          $ref: "#/components/schemas/DecimalQuantity"
        blobBaseFee:
          description: >-
            Protocol blob base fee in wei per blob gas: the EIP-4844 fee for
            excessBlobGas under the fork active at the block's timestamp.
            EIP-7918 does not change it.
          $ref: "#/components/schemas/DecimalQuantity"
        executionBaseFee:
          $ref: "#/components/schemas/DecimalQuantity"
        gasUsed:
          $ref: "#/components/schemas/DecimalQuantity"
        gasLimit:
          $ref: "#/components/schemas/DecimalQuantity"
        executionEthBurnedWei:
          $ref: "#/components/schemas/DecimalQuantity"
        blobEthBurnedWei:
          description: blobBaseFee times blobGasUsed.
          $ref: "#/components/schemas/DecimalQuantity"
        reserveFeeWei:
          description: >-
            EIP-7918 reserve price in wei per blob gas from Fusaka on, null
            before. It only changes how excess blob gas evolves and is not
            applied to blobBaseFee.
          oneOf:
            - $ref: "#/components/schemas/DecimalQuantity"
            - type: "null"
        transactionCount:
          $ref: "#/components/schemas/SafeInteger"
        targetBlobsPerBlock:
          $ref: "#/components/schemas/SafeInteger"
        maxBlobsPerBlock:
          $ref: "#/components/schemas/SafeInteger"
        transformVersion:
          $ref: "#/components/schemas/SafeInteger"
        finality:
          $ref: "#/components/schemas/Finality"
    BlobTransaction:
      type: object
      additionalProperties: false
      required:
        [network, blockNumber, blockHash, txHash, transactionIndex, senderAddress, blobVersionedHashes, blobCount, totalBurnedWei, executionBurnedWei, blobBurnedWei]
      properties:
        network:
          type: string
        blockNumber:
          $ref: "#/components/schemas/SafeInteger"
        blockHash:
          $ref: "#/components/schemas/Hash32"
        txHash:
          $ref: "#/components/schemas/Hash32"
        transactionIndex:
          $ref: "#/components/schemas/SafeInteger"
        senderAddress:
          $ref: "#/components/schemas/Address"
        blobVersionedHashes:
          type: array
          items:
            $ref: "#/components/schemas/Hash32"
        blobCount:
          $ref: "#/components/schemas/SafeInteger"
        totalBurnedWei:
          $ref: "#/components/schemas/DecimalQuantity"
        executionBurnedWei:
          $ref: "#/components/schemas/DecimalQuantity"
        blobBurnedWei:
          description: The block's blobBaseFee times this transaction's blob gas.
          $ref: "#/components/schemas/DecimalQuantity"
    BlobSchedule:
      type: object
      additionalProperties: false
      required:
        [name, activationBlock, activationTimestamp, forkId, targetBlobsPerBlock, maxBlobsPerBlock, baseFeeUpdateFraction, eip7918, source]
      properties:
        name:
          type: string
        activationBlock:
          description: >-
            Block-number label for atBlock lookups. For Dencun it is 19426589,
            the first block the blobs processor covers.
          $ref: "#/components/schemas/SafeInteger"
        activationTimestamp:
          description: The fork's parameters apply to blocks from this timestamp on.
          $ref: "#/components/schemas/SafeInteger"
        forkId:
          type: string
          pattern: "^0x[0-9a-f]{8}$"
        targetBlobsPerBlock:
          $ref: "#/components/schemas/SafeInteger"
        maxBlobsPerBlock:
          $ref: "#/components/schemas/SafeInteger"
        baseFeeUpdateFraction:
          $ref: "#/components/schemas/DecimalQuantity"
        eip7918:
          type: boolean
        source:
          type: string
          enum: [chain_spec, eth_config, configured]
...
    BlobSnapshotEntry:
      type: object
      additionalProperties: false
      required: [block, transactions]
      properties:
        block:
          $ref: "#/components/schemas/BlobBlock"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/BlobTransaction"
    BlobBlockPage:
      type: object
      additionalProperties: false
      required: [data, nextCursor, coverage]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BlobBlock"
        nextCursor:
          oneOf:
            - $ref: "#/components/schemas/Cursor"
            - type: "null"
        coverage:
          $ref: "#/components/schemas/ProcessorCoverage"
    BlobTransactionPage:
      type: object
      additionalProperties: false
      required: [data, nextCursor, coverage]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BlobTransaction"
        nextCursor:
          oneOf:
            - $ref: "#/components/schemas/Cursor"
            - type: "null"
        coverage:
          $ref: "#/components/schemas/ProcessorCoverage"
    BlobSnapshotPage:
      type: object
      additionalProperties: false
      required: [data, nextCursor, coverage]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BlobSnapshotEntry"
        nextCursor:
          oneOf:
            - $ref: "#/components/schemas/Cursor"
            - type: "null"
        coverage:
          $ref: "#/components/schemas/ProcessorCoverage"
...
    ErrorBody:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            details:
              oneOf:
                - type: object
                  additionalProperties: true
                - type: "null"
            requestId:
              type: string